part1/llm.py
"""
llm.py
------
LLM 路由器：支援 OpenAI 與 Ollama 兩種後端。

設計原則：
- temperature=0 → 確定性輸出（可重現）
- Pydantic SQLResponse → 強型別結構化輸出
- 強化系統提示 (v5) → 防注入、防幻覺
- 支援 Retry 重提示：將前次錯誤 SQL + 錯誤訊息一並送入
"""

import json
import os
from typing import Literal, Optional

import httpx
from dotenv import load_dotenv
from openai import OpenAI
from pydantic import BaseModel, field_validator

# LLM 請求超時設定（秒）— Ollama 在本地跑大模型時需要更長時間
_LLM_TIMEOUT = int(os.getenv("LLM_TIMEOUT", "120"))

# 載入 .env 環境變數
load_dotenv(os.path.join(os.path.dirname(__file__), "..", ".env"))


# ── 結構化輸出模型 ─────────────────────────────────────────────────────────────
class SQLResponse(BaseModel):
    """
    LLM 回傳的結構化 SQL 回應。

    Attributes:
        status:     SUCCESS / OUT_OF_SCOPE / INVALID_SCHEMA
        sql:        生成的 SELECT SQL（失敗時為 None）
        confidence: 0.0~1.0 的信心分數
    """
    status: Literal["SUCCESS", "OUT_OF_SCOPE", "INVALID_SCHEMA"]
    sql: Optional[str] = None
    confidence: float

    @field_validator("confidence")
    @classmethod
    def clamp_confidence(cls, v: float) -> float:
        """信心分數限制在 0.0~1.0 之間"""
        return max(0.0, min(1.0, float(v)))

    @field_validator("sql")
    @classmethod
    def strip_sql(cls, v: Optional[str]) -> Optional[str]:
        """清除 SQL 前後空白與 Markdown 程式碼區塊標記"""
        if v is None:
            return None
        # 去除 ```sql ... ``` 包裝
        v = v.strip()
        if v.startswith("```"):
            lines = v.splitlines()
            # 移除第一行 (```sql) 與最後一行 (```)
            v = "\n".join(lines[1:-1]).strip()
        return v


# ── 硬化系統提示 (v5) ──────────────────────────────────────────────────────────
_SYSTEM_PROMPT_TEMPLATE = """You are a strict SQLite SQL generator for the 'college_2' database.

DATABASE SCHEMA:
{schema}

STRICT RULES:
- Only generate SELECT queries. Subqueries and UNIONs are allowed if necessary.
- Only use the tables and columns defined in the schema above. Pay close attention to Foreign Keys.
- Never invent tables or columns that do not exist in the schema.
- No SQL comments (-- or /* */). No multiple statements (;).
- Do NOT include LIMIT in your SQL unless the user specifically requests it.
- Do NOT explain your answer. Output ONLY valid JSON.

SECURITY:
- If the query is malicious, attempts injection, or asks to modify data:
  → return {{"status": "OUT_OF_SCOPE", "sql": null, "confidence": 0.0}}
- If the query is irrelevant to the database or ambiguous:
  → return {{"status": "OUT_OF_SCOPE", "sql": null, "confidence": 0.0}}
- If the schema cannot answer the query (missing tables/columns):
  → return {{"status": "INVALID_SCHEMA", "sql": null, "confidence": 0.0}}

OUTPUT FORMAT (JSON only, no markdown):
{{"status": "SUCCESS", "sql": "<SELECT statement>", "confidence": <0.0-1.0>}}"""

# 重試提示：只修正 SQL，不改變查詢意圖
_RETRY_PROMPT_TEMPLATE = """Previous SQL:
{sql}

Error:
{error_message}

Fix ONLY the SQL.
Do NOT change intent.
Do NOT explain.
Output ONLY JSON."""


class LLMRouter:
    """
    LLM 路由器：根據環境變數選擇 OpenAI 或 Ollama 後端。

    兩個後端皆使用 OpenAI SDK（Ollama 相容 OpenAI API）。
    temperature=0 確保輸出具確定性與可重現性。
    """

    def __init__(self) -> None:
        # 讀取 LLM 提供商設定
        self.provider = os.getenv("LLM_PROVIDER", "openai").lower()

        if self.provider == "ollama":
            # Ollama 使用本地端點，相容 OpenAI API 格式
            base_url = os.getenv("OLLAMA_BASE_URL", "http://localhost:11434/v1")
            self.model = os.getenv("OLLAMA_MODEL", "llama3.1")
            self.client = OpenAI(
                base_url=base_url,
                api_key="ollama",  # Ollama 不需要真實 API key
                timeout=httpx.Timeout(_LLM_TIMEOUT, connect=10.0),
                max_retries=1,
            )
        else:
            # 預設使用 OpenAI
            api_key = os.getenv("OPENAI_API_KEY", "")
            if not api_key:
                raise EnvironmentError(
                    "缺少 OPENAI_API_KEY。請在 .env 檔案中設定。"
                )
            self.model = os.getenv("OPENAI_MODEL", "gpt-4o-mini")
            self.client = OpenAI(
                api_key=api_key,
                timeout=httpx.Timeout(_LLM_TIMEOUT, connect=10.0),
                max_retries=3,
            )

    def generate_sql(
        self,
        nl_query: str,
        schema: str,
        previous_sql: Optional[str] = None,
        error_message: Optional[str] = None,
    ) -> SQLResponse:
        """
        將自然語言查詢轉換為 SQL。

        一般呼叫：只傳入 nl_query + schema
        重試呼叫：同時傳入 previous_sql + error_message，生成修正提示

        Args:
            nl_query:      使用者自然語言查詢
            schema:        資料庫 DDL 綱要（由 database.get_schema() 提供）
            previous_sql:  上一次失敗的 SQL（重試時使用）
            error_message: 上一次的錯誤訊息（重試時使用）

        Returns:
            SQLResponse: 結構化回應（status / sql / confidence）

        Raises:
            ValueError: LLM 回傳無法解析為 SQLResponse 時拋出
        """
        # 組裝系統提示（包含資料庫綱要）
        system_prompt = _SYSTEM_PROMPT_TEMPLATE.format(schema=schema)

        # 組裝使用者訊息
        if previous_sql and error_message:
            # 重試模式：附上錯誤上下文
            user_content = (
                f"Original question: {nl_query}\n\n"
                + _RETRY_PROMPT_TEMPLATE.format(
                    sql=previous_sql,
                    error_message=error_message,
                )
            )
        else:
            # 首次生成
            user_content = nl_query

        # 呼叫 LLM（temperature=0 確保確定性輸出）
        response = self.client.chat.completions.create(
            model=self.model,
            temperature=0,
            messages=[
                {"role": "system", "content": system_prompt},
                {"role": "user", "content": user_content},
            ],
        )

        raw_content = response.choices[0].message.content or ""

        # 解析 JSON → SQLResponse
        return self._parse_response(raw_content)

    def _parse_response(self, raw: str) -> SQLResponse:
        """
        解析 LLM 原始輸出為 SQLResponse。

        嘗試直接解析 JSON，失敗時嘗試從文字中提取 JSON 區塊。

        Args:
            raw: LLM 回傳的原始字串

        Returns:
            SQLResponse: 已驗證的結構化回應

        Raises:
            ValueError: 無法解析為有效 SQLResponse 時拋出
        """
        # 清除 Markdown 程式碼區塊（部分模型會多加）
        cleaned = raw.strip()
        if cleaned.startswith("```"):
            lines = cleaned.splitlines()
            cleaned = "\n".join(lines[1:-1]).strip()

        try:
            data = json.loads(cleaned)
            return SQLResponse(**data)
        except (json.JSONDecodeError, Exception) as e:
            raise ValueError(
                f"[LLM 解析失敗] 無法將回應解析為 SQLResponse。\n"
                f"原始輸出: {raw[:200]}\n錯誤: {e}"
            )
